added getStepsizes
getStepsizes is a method for the msdanalyzer object that returns a cell
array with the step sizes, either in x and y or absolute distance, for
each step in each track.
Inspecting the distribution of step sizes can be useful to show the
precision (lower step size), find different diffusion speeds (multiple
peaks), comparing between experiments with different time steps (do step
sizes get bigger, or are they limited by precision)
Code is based on "getVelocities".