added getSDdt
getSDdt is a method for the msdanalyzer object that returns a cell array
with, for each step in each track, the squared displacement for that
step divided by the time step.
This complements the "getStepsizes" function: the distribution of the
square displacement per time should be the same between experiments with
different time steps. Different peaks can indicate different diffusive
states of the tracked particles.
Code is based on "getVelocities" and "getStepsizes".